+++ game-bellwright/CHANGELOG.md
@@ -1,9 +1,16 @@
 # Changelog
 
 ## Future Improvements (Not Yet Released)
-- 
+
+- Added tool to launch ModKit (Epic Games).
+
+## [0.3.0] - 2025-06-26
+
+- Disabled Signature Bypass downloader since it is no longer required with the ModKit release.
+- Changed Pak mod installer for new ModKit format, moved installation folder to "Content/Mods", and disabled Load Order (now in-game).
 
 ## [0.2.0] - 2025-04-25
+
 - Added support IO Store pak mods (pak/ucas/utoc)
 - Updated signature bypass download to new mod page
 - Added buttons to open multiple folders - folder icon on Mods toolbar